refactor: (tobe) 메뉴 패키지 구조 리팩터링
