--- src/main/java/kitchenpos/menus/domain/tobe/domain/menu/JpaMenuRepository.java
@@ -1,11 +1,0 @@
-package kitchenpos.menus.domain.tobe.domain.menu;
-
-import kitchenpos.common.domain.MenuId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, MenuId> {
-
-    // @Query("select m from Menu m, MenuProduct mp where mp.productId = :productId")
-    // @Override
-    // List<Menu> findAllByProductId(@Param("productId") ProductId productId);
-}
