--- src/main/java/kitchenpos/menus/domain/tobe/domain/menu/MenuProductSeq.java
@@ -1,42 +1,0 @@
-package kitchenpos.menus.domain.tobe.domain.menu;
-
-import java.io.Serializable;
-import java.util.Objects;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Embeddable
-public class MenuProductSeq implements Serializable {
-
-    @Column(name = "seq")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    private Long value;
-
-    protected MenuProductSeq() {
-    }
-
-    public MenuProductSeq(final Long value) {
-        this.value = value;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final MenuProductSeq that = (MenuProductSeq) o;
-        return Objects.equals(value, that.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-}
