--- src/main/java/kitchenpos/menus/domain/tobe/domain/menugroup/JpaMenuGroupRepository.java
@@ -1,7 +1,0 @@
-package kitchenpos.menus.domain.tobe.domain.menugroup;
-
-import kitchenpos.common.domain.MenuGroupId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, MenuGroupId> {
-}
