+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/JpaMenuGroupRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.menus.tobe.domain.menugroup;
+
+import kitchenpos.common.domain.MenuGroupId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaMenuGroupRepository extends MenuGroupRepository,
+    JpaRepository<MenuGroup, MenuGroupId> {
+
+}
